[cleanup] Extract most diagram-specific conversion code into a seperate package
